+++ winpr/include/winpr/assert.h
@@ -35,7 +35,7 @@
 		{                                                                                     \
 			wLog* _log_cached_ptr = WLog_Get("com.freerdp.winpr.assert");                     \
 			WLog_Print(_log_cached_ptr, WLOG_FATAL, "%s [%s:%s:%" PRIuz "]", #cond, __FILE__, \
-			           __FUNCTION__, __LINE__);                                               \
+			           __FUNCTION__, (size_t)__LINE__);                                       \
 			winpr_log_backtrace_ex(_log_cached_ptr, WLOG_FATAL, 20);                          \
 			abort();                                                                          \
 		}                                                                                     \
